chore: move helper folder content to lib
