+++ lib/server/api/error-handler.ts
